requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
